Add mtconnect_adapter_sdk as an explicit dll callout
By explicitly calling out the sdk's dll via a relative path we hope to avoid triggering a rebuild.

The MSBuild@1 task does not really support platform and configuration. The platform has defined Debug as default in the common region of the csproj so it is overridden using the msbuildArguments parameter