Specify output location for pack
Without specifying the output directory as $(Build.ArtifactStagingDirectory) the package was going to the bin/Release folder making it not find-able by the push step.

The real solution here is to  make a PR to https://github.com/microsoft/azure-pipelines-tasks that supports version-suffix (and similar) in the task so a custom dotnet command is not required.