Remove symbol generation and version include githash
Packaging the source and symbols in the nupkg was generating a second package which conflicted on push to the artifact feed. This removes that generation.
In the future we should try to release symbols using the PublishSymbols@2 task (https://docs.microsoft.com/en-us/azure/devops/pipelines/tasks/build/index-sources-publish-symbols). For this story setting up the Artifact symbol storage was outside of scope.

Changed the version to indicate generation from a PR and include the githash for the build for simplicity.

Updated readme with build status badge and title rename.