Explicit triggers. Add nuget build reasons. Logs
A commit to develop after a PR was closed did not trigger a CI build. This is an attempt to be more explicit for both the develop branch and the master branch.

Triggering off a manual build does not indicate the branch that it came from, rather it does a detached HEAD which means it does not have a branch specifier. To correct this we catch if it is a Manual build reason.

To handle the inevitable unknown values that come from not having a local-installable version of Azure Pipelines we now complete the NuGet version suffix with defined unknown naming scheme.

Add logging to print out build variables of interest at the top of the job. This will be useful for creating a template build job in the future.